InnerClass Structure added to ClassInfo
