+++ typescript/src/main/java/org/hisrc/jsonix/Schema/XSD/AnySimpleType.java
@@ -1,0 +1,8 @@
+package org.hisrc.jsonix.Schema.XSD;
+
+import org.hisrc.jsonix.Model.TypeInfo;
+
+public class AnySimpleType extends TypeInfo
+{
+
+}
